refactor: some minor adjustments about dfget
